fix: Render multiline labels in legend and tooltip
`chart.js@2.8.0` has a bug where multi line labels are not displayed
properly in tooltips and in the legend. They are displayed by using
`Array.toString` so the result is a comma joined string of the words.
There is a `chart.js` pull-request
([#6117](https://github.com/chartjs/Chart.js/pull/6117))
which resolves this, but until it gets merged, we fix it in
`ng2-charts` by monkey-patching `chart.js`.

This commit provides two public api functions:
* monkeyPatchChartJsLegend()
* monkeyPatchChartJsTooltip()

App developers may call either or both to enable this feature.